+++ src/ohMyForm_analyze/report.nim
@@ -1,22 +1,43 @@
 from std/tables import Table, `[]`
 from std/json import items, `{}`, JsonNode, len
-from std/strutils import multiReplace
+from std/strformat import fmt
+from std/strutils import multiReplace, parseInt
 from std/uri import decodeUrl
 import karax / [karaxdsl, vdom]
 
 import ohMyForm_analyze/core
 
-proc createReport*(scores: Table[string, int]; entry: JsonNode): string =
+proc createReport*(scores: Table[string, string]; entry: JsonNode): string =
+  var
+    total = 0
+    numFields = 0
   let vnode = buildHtml(table):
     tr:
       th: text "Name"
       th: text "Value"
       th: text "Score"
     for field in entry{"fields"}:
-      tr:
-        th: text field{"field", "title"}.getStr
-        th: text field{"value"}.getStr
-        th: text $scores.getScore field.getId
+      try:
+        let score = scores.getScore field.getId
+        total += score.parseInt
+        inc numFields
+        tr:
+          th: text field{"field", "title"}.getStr
+          th: text field{"value"}.getStr
+          th: text score
+      except ValueError:
+        discard
+    let nflds = numFields / 10
+    h1: text fmt"Total: {total} ÷ {nflds} = {total.float / nflds}%"
+        
+
+  proc clean(s: string): string =
+    s.multiReplace({
+      "</#text>": "",
+      "<#text>": "",
+      "\n": "",
+      "  ": " "
+    })
   result = """<style>
 table {
   border-collapse: collapse;
@@ -30,11 +51,6 @@ th, td {
 
 tr:nth-child(even) {background-color: #f2f2f2;}
 </style>
-"""
-  result.add decodeUrl ($($vnode)).multiReplace({
-    "</#text>": "",
-    "<#text>": "",
-    "\n": "",
-    "  ": " "
-  })
+""".clean
+  result.add decodeUrl ($($vnode)).clean
   echo result