+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Version 1.0.0 (15/03/2023)
+
+- Added report generation
+- Added single entry analysis
+- Added new UI to data
+- Added way to share the page
+
+---
+
 ## Version 0.1.0 (15/03/2023)
 
 - Added mocking to get data and prevent page drawing